0818を正としてメイン開発フォルダを差し替え、旧構成をold_0808/へ退避
0818/librim/RIManager/の実装を、リポジトリのメイン開発フォルダ
(core/products/include/test/docs/samples/CMakeLists.txt)として採用した。
0818フォルダ自体は読み取り専用のオリジナルとしてそのまま残す。

直前までのメイン開発フォルダ(0808ベースの実装。今回のセッションで行った
AdapterDispatcher::Dispatch(RIDataId, RIMValue)化やドキュメント一式を含む)
は old_0808/ へ退避。0818側の設計はAdapter/Publisher/Storageまわりが
大きく再構成されており(printer_a専用ID/型ヘッダの導入、
RouteResolver/DomainStorage/AccessorFacadeの追加等)、0808の実装とは
別物として両方の履歴を残す。

README.md/mainfw_mock/.gitignore/デバッグ環境構築手順.md は0818側に
対応物が無いため変更していない。

+++ core/common/include/DeviceEvent.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "rim_data_id.h"
+#include "RIId.hpp"
 
 namespace rim
 {
@@ -9,7 +9,7 @@ struct DeviceEvent
 {
     RIDataId id;
 
-    double value;
+    int value;
 };
 
 } // namespace rim
